Semantic updates, add a generic map, other fixes
Started working on departing from C conversion rules, which are awfully broken.
What I do now is, based on the declaration type if any, to expand types of expressions based on a new set of rules.
So now literal values aren't strictly a certain size but it depends whether it can be expanded to the declaration type.
In the case of `var` type inference, integers are `int32` by default, while decimals `float`.
Let's see how it goes...