Add `typeof`, `sizeof`, `alignof`, `offsetof`
The big issues with these is of how they work in C. Well, not really, as this is not C, but I'm adding the same keywords.
I need to know before-hand if I'm parsing an expression or a type. Since it's much used the kind `sizeof(T)`, I opted to only allow types for `sizeof` and `alignof`.
If one really needs to get the type of an expression, just do `sizeof(typeof(expr))`.
A bit verbose, but not that bad.
`typeof` allows you to use as type what type the expression evaluates into.