`#staticassert`, `#notes` syntax, error line print
`#notes` are supposed to be compile-time directives one can use.
Thus, they can be use to add hints to the compiler.
This includes `#staticassert`, which asserts a constant expression at compile time.
Also made sure casts to constant values generates constants that can be folded correctly.
I'm still not entirely sure whether to allow notes to be applied to declarations or not, because it may be ambiguous.
Maybe I'll fix it soon.