Fix line offset
Incrementing line offset at the `lex_scan` isn't enough.
Doing it at `lex_next` is perfect.
Also fixed the loc position for expressions.
To work correctly with binary/unary and other operators, we need to take the offset for when the character is at the operator position.
Should be the right position, hopefully.