Attempt to fix scanning of big `u/int64` integers
This probably still needs a check.
All integer literals start out as unsigned, so that we can represent a whole `uint64`.
Though, right now it's signed.
Working with the biggest unsigned int should work anyway.
And it becomes signed right when a `-` is added.
I just need to understand how to nicely add checks for overflow.
For instance, negative the max uint64 is slightly questioning.