Add overflow checks when constant folding ints
Every integer literal starts out as unsigned.
To be precise `uint64`, which allows us to represent the whole range of integers we are interested about.
I can check based on the type given to the int if it overflows its maximum type size.
Same thing for constant folding of binary operations and applying `-` on an int too large for its signed type.
I'm not entirely sure what other overflows other than `+` `-` `*` can happen.
`INT_MIN / -1` in theory overflows. I also overflow if a left shift is too large for the bitsize of the type.
Should be good enough.
Later cases can be patched/fixed.