Fix overflow check, constant fold bools `not true`
So, to get correct results when using `__builtin_*_overflow` builtins, it's better to cast everything.
I'm glad I'm now aware of this, because now it correctly works for underflows.
So now for example you can define the min `int32` as: `-2147483647 - 1` because otherwise it gives an error because it tries to apply the `-` on `2147483648` (overflow).
Well, we know that it doesn't overflow, but whatever.
Even C headers define these like so, no big deal.
To finish off constant folds, allow folding of `not not true` or `false`.