Integer suffixes and add untyped integers
Normal integer literals when there is no context, which is when there is type inference and in expression statements default to `int32`.
However, this doesn't work nicely in certain cases.
Thus, integer literals may be "untyped", meaning they can be widened when used in unary and binary operations, but also when folding the AST.
Hopefully I didn't break previous cases.
Also added suffixes `i32`, `i64` and so on.