Begin main regalloc and live analysis logic, fixes
Improved again IR and code generation.
Renamed some files.
Some pointers from the AST could be made so that you can use for both type specs and real types.

Added insert/remove_at functions for vectors. Gotta be careful with those macros!

And most importantly, started working on the main register allocation logic. Just a linear scan register allocator.
While register allocation is supposed to be an optimization, there are various constraints from the ABI, x64 itself and calling conventions that expect stuff in specific registers.

The nice thing of the register allocator is that all it'd need is just the total number of physical registers + the number of temporary registers that can be used. There is some mapping to be done for function parameters and the `IR_CALL` and `IR_PUSHARG` have to setup the indexes correctly.

The idea would be to codegen -> regalloc -> emit. There are some details to fill out about offsets of stack variables and living ranges. And the fact that many instructions in x64 expect 2 operands.
I love you Intel.

Either inserting new IR instructions or handle it directly in the emitting process.
There's an intermediate phase called "instruction selection", where you "lower" the IR into a format that allows to do pattern matching and generate the right instruction.

`vp_sel_x64` for instruction selection, but we just directly emit x64. We can later expand it, but it'd also require changes in the regalloc. It's all interconnected!

I'm just hard-coding Windows x64 specifics for the time being.