Basic epilogue/prologue, call patching, fixes
It's better to not use AST declarations in "selection" and emitting, as at that point of the pipeline the AST itself could be entirely discarded.

Instead, I have this `Code` struct to represent the various function "code blocks" and the infos needed for handling epilogue and prologue + stack. The only units we are interested in for generating x64 code are only functions with bodies, that's it. No need to carry around the AST.

Some details missing, like having to save and restore some registers before and after a call, and actually following Windows x64 ABI more carefully. (very stupid details)

And especially handling reference taken vregs, and assign stack offsets in case of a compound literal (in some cases).

Call patching is used to generate correct addresses for private calls, which is basically the same way it's handled for jumps...
So you can now call functions, yay.

The main goal now I guess would be to somehow get to interact with OS functions, print on screen.

That'll take a bit...