Handle `&` taken vars, field of pointer to struct
In the process of supporting code generation for address taken variables, also fixed the spilling of vregs, which is used in this case.

Well, you can't take the address of a register. Just need to be handled for local/global variables.

Also added the direct case of a compound literal being address taken `&{a = 23, b = false}`