Rework `memcpy` and `memzero`, use `vec_t`, fixes
Having a single `IR_MEMCPY` and `IR_MEMZERO` for handling structs/unions while ideally simple turns out to be a major hassle.

A simple mem copy or mem set in x86 is kind of simple using the `rep` prefix with string instructions, but it might not be simple with other ISAs, I think.

Besides, it's not very robust and there are a million ways to efficiently initialize or copying structs.

The gen `memcpy`/`memzero` emit a pretty chuky block of code, but it's ok for now. The generated code overall isn't very efficient.

Fixed stuff to use the `vec_t` for a bit more visual clarity. AST nodes could be changed to not use any sort of vectors, they're basically immutable.

And finally, turns out that tracking in/out register usage for each basic block are a must. This is because otherwise for branching code the regalloc might end up wrongly re-using registers, thus corrupting the program.

Yikes...