Basic load of float constants
The basic functioning is same as the handling of imm64 in case of non-MOV IRs, which is a limitation of x86/64.

You can't directly load an immediate into an XMM register! So it needs to either be loaded from memory (data section of the executable) or directly moving the immediate representing the float/double into a general purpose register and then moving that register into an XMM one with `movq` or `movd`.

Not sure which one, `movq` is supposed to be for quad words, while `movd` for double words? gcc MinGW seems to prefer `movq`, who knows.

Even weird detail, gcc on Linux always generates a load from memory to set the float constant into the register.

I still don't really get it why. Maybe to save some bytes? 2 loads are always generated no matter what, because of the gpr to xmm step.