Optimization pass
It's not really an optimization pass, as right now it just removes unused vregs, basic blocks, and peephole optimizations.

The middle end of compilers tends to become very huge, and there is a lot of matieral about it.

The main optimization which could later be added is SSA, which accounts for most of the optimization transformations.
It's basically the go-to representation for optimizations!

And there is much more left around instruction selection. x86 is notoriusly complex and some encodings are shorter than others. For example using an LEA for some additions or shifts instead of integer power of 2 multiplications/divisions.

It could be done with the current IR?
Not really? While I still end up having to "tweak" the IR to account for destructive x86 2-operands instructions and float constants, the IR as is can't really represent other more complex modes for x86.
And these pretty much only exist in x86! ARM and RISC-V are much straightforward, I think.

Also a small fix to sel_ret.
And a fix to use a basic block for the returns, is it necessary? I guess it keeps the rule of a non-empty basic block ending with a jump.