Compound assignment
Introducing parsing, semantic check, and code generation for compound assignments, which are +=, -=, *= and so on.

The naming is a bit confusing, is there another way to call them?
Code generation is pretty straightforward, except when the left handside is more complex.

Since I want to follow C style rules, left handside must be evaluated only once.
I guess it's ok?