`while` loop, `break` and `continue`
Adding the usual loop constructs.

Next up would be `for` loop, or rather firstly adding `++` and `--`.

They're simple to add, the actual tricky stuff is beyond IR generation, in the middle-end and backend.

As one may know, the middle-end of a compiler can become very huge and intricate!
That's because frontend is a solved problem, while generating optimal code isn't, and may never be entirely :)

There's always an optimization missing.