`intcast`, `floatcast`, `ptrcast` and `bitcast`
And some fixes!

These are different kinds of casts, ala Zig. But unlike Zig, you have to give it the type for the cast.

I still have to test and see what's missing to fix. `IR_CAST` ends up being complex. x64 and cast semantics have some quirks for combinations of casts.

For instance, what would casting a `uint64` or `int64` to `double` should even give? In C it's easy: undefined.
I... leave it to be implemented later.

In actuality, a cast should be emitted for every mismatch between 2 types, which I still have to add a check for...
Some implicit casts still end up requiring a `IR_CAST`.