`[[attribute]]`s, PE imports, strings, fixes
Pretty much changes everywhere.

Added `[[attribute]]` syntax for decorating functions. The first attribute added is `[[lib("library")]]`, which is used to specify from which external library a function prototype comes from.
Other attributes can be added later on.

Since right now the compiler only targets Windows PEs, we're required to add a new section `.idata` for handling DLL functions. It's always generated with 2 functions from `kernel32.dll`...

Well, we can already print something on the terminal! _nice_

The way the PE is generated is not ideal. An actual compiler would just emit an object file (COFF for Windows) with relocations and the linker would take care of stitching everything together, including external imports.

I guess we'll need to emit object files. But the idea for this compiler would be to provide everything under a single executable.

Well, others already do so, and under the hood invoke the linker and what else...

Besides, added support for `var s : uint8[] = "a string literal"`, which becomes an array of chars allocated on the stack.
Support for `var s : uint8*` requires `.data` section for Windows.

Somehow handled functions passed as pointers to variables or other locations. It requires to patch `lea reg, [rip +/- ofs]` with the address of the function in the `.text` section,
And, it seems to work.

There's a lot of patching currently involved, and more will be required for handling globals and static strings.

Fixed some stack handling, mainly for function arguments, since `WriteConsoleA` require a stack parameter. Almost seems made on purpose...

Previous examples still work, so that's enough.