Re-structure files, `isize`/`usize`, `float32/64`
Re-ordered a bit some of the backend file.

There isn't any actual instruction selection, so now there is:
- `vp_ir_x64`
- `vp_low_x64`
- `vp_link_x64`

And their respective "interfaces". It's going to be more interesting with more archs added.

Also re-fitted better for handling targets, since this is supposed to be a cross compiler.

Linking right now is just the current patching logic. Handling for object files (COFF on Windows) to be done.

Oh yea, also added `isize` and `usize`, while also renaming `float` and `double` to `float32` and `float64`.
And for `isize` and `usize` the literal suffixes `uz` and `iz`.

I'm not totally sure if `sizeof`, `alignof`, `offsetof` and other friends should use `usize` for the type.

Maybe later?
Casting logic should remain the same.