Add `alias` declaration
As of right now, `alias` and `type` are exactly the same thing.

But the idea would be to use `alias` as a C style `typedef`, ie an alias, and `type` for "stronger types".

Types "created" using `type` won't be implicitly converted, they would require implicit casting. That's it, that's the idea.
It's nice to have in a few cases.