Fix parsing of typed compound literals
Turns out, `Foo {...}` can be ambiguous when for example in an `if` condition.

We could require a bunch of parens, but it's better to switch the syntax into `Foo::{...}`.

More semantics added to the `::` "operator", but it shouldn't be too much.

Reserved 2 more keywords for later implementation.

The System-V ABI specification for varargs arguments infuriates me, so we'll ignore those for the time being.