`import`, `pub` and module system, fixes
Adding support for imports and a module system required some changes to parser (of course) and semantic analyzer.

Each import is ran only once, which is based on the absolute path of the file.
Only the function bodies from the "root" module, which is the target file for compilation into object file, are semantically checked.

Of course you can only import `pub` prefixed declarations from a module, and either some specific symbol, or everything, `*`.

Types, aliases, defs and others need to be semantically checked regardless of the module context.

There's still some details to add, like the use of `::` for module access. You can already create a _module_ symbol from an `import` declaration, but it's missing this.

There's also the question of name mangling, not really mangling, but rather how to handle function symbols nicely by prefixing each with the name of the module they're in, example: `vp::sema::resolve` or `sema::resolve`.