Refactor x64 emit functions
The emit functions for x64 were pretty much a mess. Basically now you represent every kind of x64 register and size using a single value representation, `X64Reg`. Another representation is provided for memory operands, `X64Mem`.

This is useful, as it's supposed to be a single interface for accessing x64 registers and emit functions in IR lowering, asm instructions and IR to x64.
Some macros and helpers could be added for improving readability, as it's pretty "hairy".

How to do instruction selection? It's pretty much a required step for generating efficient x64 instruction encodings. It's not quite needed for ARM64 and RISC-V64.

I guess we could start and add ARM64 support for emit functions. After all, this is supposed to be a cross-platform compiler.

x64 asm support is to be fixed later.
We also need a way for somehow supporting the direct use of named variables inside `asm` blocks. But how!

Though, there's a whole sleuth of other stuff in the compiler to improve upon...