Improve vecs and arenas
While it's nice that a `NULL` vec is treated as an empty vector, it's not nice the additional checks needed each time a corresponding macro is used.

So now, vecs have to be first initialized, using `vec_init`. Maybe a similar macro with a chosen count of elements can be added later on.
Least to say, vecs are pretty much used everywhere in the codebase.

Lastly, arenas do not use a vec internally for storing the blocks of memory. Switched to a linked list (of blocks). Should be good enough.