`const` variable declarations
`const` variables implemented.

In the semantic checker they're treated as a separate symbol, which is then used to do the error check on name assignment.

Basically the language right now has 2 concepts of `const`-ness, a variable slot being `const` and/or the underlying type.

It's a bit harder to enforce const-ness of the whole types. Or maybe not?